feat: 예약 완료 페이지에서 클래스명 출력 #44
