+++ src/app/api/supabase/fetchClassInfo.ts
@@ -2,11 +2,11 @@ import { PostgrestSingleResponse } from '@supabase/supabase-js';
 import { supabase } from './supabase';
 import { ReserveClassType } from '@/types/class';
 
-// 메인페이지, 리스트페이지, 디테일페이지, 예약페이지 유저 정보 불러오는 함수
-export const fetchClassInfo = async ({ classId }: { classId: string }) => {
+export const fetchReserveClassInfo = async ({ classId }: { classId: string }) => {
+  console.log(classId);
   const { data: classInfo, error }: PostgrestSingleResponse<ReserveClassType> = await supabase
     .from('class')
-    .select('*')
+    .select('class_id, category, title, location, price, image, max_ppl')
     .eq('class_id', classId)
     .single();
 
@@ -16,4 +16,17 @@ export const fetchClassInfo = async ({ classId }: { classId: string }) => {
   }
 
   return classInfo;
+};
+
+// 여러 클래스 정보를 불러오는 함수
+//따로 만든 이유 -> single이 단일 정보만 불러오기 때문
+export const fetchClassInfos = async () => {
+  const { data: classInfos, error } = await supabase.from('class').select('*');
+
+  if (error) {
+    console.error('클래스 정보들 불러오기 오류 => ', error);
+    return [];
+  }
+
+  return classInfos;
 };
