+++ src/app/reserve/page.tsx
@@ -2,11 +2,11 @@ import ClassInfo from '@/components/reserve/ClassInfo';
 import DateTimePicker from '@/components/reserve/DateTimePicker';
 import PriceCalculator from '@/components/reserve/PriceCalculator';
 import ReserveButton from '@/components/reserve/ReserveButton';
-import { fetchClassInfo } from '../api/supabase/fetchClassInfo';
+import { fetchReserveClassInfo } from '../api/supabase/fetchClassInfo';
 
 export default async function ReservePage({ searchParams }: { searchParams: { classId: string } }) {
   const classId = searchParams.classId;
-  const classInfo = await fetchClassInfo({ classId });
+  const classInfo = await fetchReserveClassInfo({ classId });
 
   return (
     <div className="w-full h-full">
